IDP_India/Modularized/requirements.txt
streamlit
base64
time
PyMuPDF
pickle
re
os
json
hashlib
datetime
pathlib
typing
logging
opencv-python
numpy
Pillow
io
openai
langchain
langchain-openai
langchain-community
langchain-core
requests
python-dotenv
pandas
openpyxl
psycopg2-binary
smtplib
email
